+++ r_main.c
@@ -444,7 +444,7 @@ void R_InitPointToAngle (void)
     // UNUSED - now getting from tables.c
 #if 0
     int	i;
-    long	t;
+    int	t;
     float	f;
 //
 // slope (tangent) to angle lookup
